% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHeatmap.R
\name{AnnotationHeatmap}
\alias{AnnotationHeatmap}
\title{AnnotationHeatmap}
\usage{
AnnotationHeatmap(peaks, peaknames = "mypeaks", annotation,
  annoname = "annotation", span = 2025, step = 50,
  plotcols = circlize::colorRamp2(c(0, 5), c("white", "darkblue")),
  plotHM = TRUE)
}
\arguments{
\item{peaks}{A Granges object containing your ChIP peaks. Must include seqnames (chromosomes), start, end, strand, summit, and name.}

\item{peaknames}{A character string that describes your peaks. Default is "mypeaks".}

\item{annotation}{A GRanges object containing the annotation ranges you want to plot around peak summits.For example, promoter regions.}

\item{annoname}{A name to describe the annotationthat was provided (for example: "promoters"). Defaults to "annotation".}

\item{span}{The distance from the peak summit to the left and right that you want your heatmap to extand to. Default is 2025.}

\item{step}{The window size in which reads are counted/plotted in the heatmap. Default is 50.}

\item{plotcols}{The colors for the heatmap. Default is circlize::colorRamp2(c(0, 5), c("white", "darkblue")).}

\item{plotHM}{Also plot the heatmap or only generate overlap counts in the windows? Default is TRUE.}
}
\value{
A matrix that contains the overlap with the annotation in each window (column headers=middle of the window)
for each peak (rownames=peaknames). 1 = at least half of the window overlaps with a region in the annotation
object, 0 = less than half of the window overlaps with the annotation object. If plotHM is TRUE, a
heatmap is plotted as well and saved in a directory (in your current working directory. if it already exists,
it is not overwritten) that is named after the peaknames object. This heatmaps is clustered.
}
\description{
This function generates a heatmap of Genomic Ranges around the summits of ChIP peaks.
}
\details{
This function generates a heatmap of annotations around the summits of ChIP peaks. An annotation GRanges object must be provided.
This can contain regions representing anything of interest, for example promoters, genes, repeats,
or transcription factor motif occurences in the genome. If at least oen region in the annotation GRanges object overlaps a window
in the heatmap by 1/2 of the step size provided, this window will be called overlapping (= 1), otherwise it will be called non-overlapping (= 0).
}
\examples{
peaks <- GenomicRanges::GRanges(
seqnames = Rle(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
ranges = IRanges(50101:50110, end = 51111:51120),
strand = Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
summit = 1:10, name = head(letters, 10))
annotation <- GenomicRanges::GRanges(
seqnames = Rle(c("chr1", "chr4", "chr1", "chr3"), c(1, 3, 2, 4)),
ranges = IRanges(50101:50110, end = 51111:51120),
strand = Rle(strand(c("-", "+", "*", "-", "-")), c(1, 2, 2, 3, 2)),
name = head(letters, 10))
AnnotationHeatmap(peaks=peaks,annotation=annotation,plotHM=FALSE)

}
