% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTracks.R
\name{calcTracks}
\alias{calcTracks}
\title{calcTracks}
\usage{
calcTracks(
  bigwigFiles,
  bigwigNames,
  bigwigGroupNames,
  ROI,
  CoverageSmoothing = TRUE,
  smoothing_window = 20,
  reps,
  genes,
  exons
)
}
\arguments{
\item{bigwigFiles}{Character vector containing the filenames (including the full path) of read alignment files in bigwig format.}

\item{bigwigNames}{Character vector containing the names to describe the \code{bigwigFiles} you are using (for example: "H3K9me3_replicate1").}

\item{bigwigGroupNames}{Character vector containing the group names to describe the \code{bigwigFiles} you are using (for example: "H3K9me3").}

\item{ROI}{A GRanges object containing your single region of interest that should be plotted. Must include seqnames (chromosomes), start, end.}

\item{CoverageSmoothing}{Logical scalar, indicating whether the coverage tracks should be smoothed. Default: TRUE.}

\item{smoothing_window}{Numeric scalar providing the window size for smoothing. Default: 20.}

\item{reps}{A GRanges object containing all repeats in the genome, or at least the ones overlapping the \code{ROI}.}

\item{genes}{A GRanges object containing all genes in the genome, or at least the ones overlapping the \code{ROI}.}

\item{exons}{A GRanges object containing all exons in the genome, or at least the ones overlapping the \code{ROI}.}
}
\value{
A list of data frames containing all necessary information to plot
the screenshots using the \code{plotTracks} function.
}
\description{
Prepare ChIP, repeat, and gene annotation data for plotting tracks
for a given region of interest (ROI).
}
\details{
calcTracks is used to prepare ChIP, repeat, and gene annotation data for plotting tracks
for a given region of interest (ROI).
}
\examples{
#see vignette

}
