% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativePlots.R
\name{CumulativePlots}
\alias{CumulativePlots}
\title{CumulativePlots}
\usage{
CumulativePlots(
  counts,
  bamNames,
  span = 2025,
  step = 50,
  summarizing = "mean",
  overlapNames = "NA",
  plot = TRUE,
  confInterval = 0.95,
  plotcols = c("violet", "darkgrey"),
  overlapLabels = c("overlap", "no overlap")
)
}
\arguments{
\item{counts}{A list of counts across peak summits generated by SummitHeatmap.}

\item{bamNames}{A character vector of names to describe the bam files you are using. This should match colnames(counts).}

\item{span}{The distance from the peak summit to the left and right that you want your heatmap to extand to. Default is 2025.}

\item{step}{The window size in which reads are counted/plotted in the heatmap. Default is 50.}

\item{summarizing}{The function to summarize the heatmaps across peaks to achieve a cumulative line. Default is "mean".}

\item{overlapNames}{A character vector of the row names of the counts that overlap with a given annotation, for example, transcription start sites.}

\item{plot}{If TRUE (default), returns the cumulative plots, otherwise returns the underlying data table.}

\item{confInterval}{The confidence interval around the mean that will be calculated and plotted. Default is .95. Only used if summarizing = "mean".}

\item{plotcols}{Two different colors for the lines that contain the values for the peaks defined in overlapNames, and the rest of the peaks.}

\item{overlapLabels}{Two different labels for the lines that contain the values for the peaks defined in overlapNames, and the rest of the peaks.}
}
\value{
Returns a matrix with mean values of the log2 of supplied counts (with pseudocount of 1 added) in each bin for each sample, split by overlap_names.
Plots of mean values for read counts around position of interest in the genome.
One plot for each count matrix supplied. All plots are arranged on one page and saved as a png.
}
\description{
Summarizes and plots the values for counts around positions of interest in the genome.
}
\details{
This function plots summarized (default = mean) values of pre-calculated read counts around the positions of interest in the genome, for example summits of ChIP peaks.
}
\examples{
counts <- list(matrix(rnorm(21000,2,1),ncol=81,nrow=100,dimnames=list(1:100,-40:40)),
               matrix(rnorm(21000,2,1),ncol=81,nrow=100,dimnames=list(1:100,-40:40)))
bamNames <- c("counts1","counts2")
names(counts) <- bamNames
CumulativePlots(counts,bamNames=bamNames,overlapNames="NA")

}
