% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMotifs.R
\name{FindMotifs}
\alias{FindMotifs}
\title{FindMotis}
\usage{
FindMotifs(
  peaks,
  peaknames = "mypeaks",
  topn = 100,
  width = 100,
  runMEME = TRUE,
  runHOMER = TRUE
)
}
\arguments{
\item{peaks}{A Granges object containing your positions of interest. Must include seqnames (chromosomes), start, end, unique peak (row) name.}

\item{peaknames}{A character string that describes your peaks. Default is "mypeaks".}

\item{topn}{The number of top peaks (counted from the top row of the GRanges object) to use for motif finding. Default is 100.}

\item{width}{The width of the peak (around the center of each range) to use for motif finding. Default is 100.}

\item{runMEME}{Should MEME motif finding be run? Default = TRUE.}

\item{runHOMER}{Should HOMER motif finding be run? Default = TRUE.}
}
\value{
Saves the outputs of MEME and HOMER.
}
\description{
Finds enriched motifs in supplied regions (peaks) using MEME and HOMER.
}
\details{
Finds enriched motifs in supplied regions (peaks) in the mouse mm10 genome using MEME and HOMER. It will write a fasta file with the peak sequences
(for MEME Input) and a bed file with the peak coordinates (for HOMER Input). If runMEME/runHOMER is set to TRUE (the default)
it will also run the MEME/HOMER analysis (requires MEME/HOMER to be installed).
}
\examples{
peaks <- GenomicRanges::GRanges(
seqnames = Rle(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
ranges = IRanges(50101:50110, end = 51111:51120),
strand = Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
summit = 1:10, name = head(letters, 10))
FindMotifs(peaks)

}
