% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCbias.R
\name{GCbias}
\alias{GCbias}
\title{GCbias}
\usage{
GCbias(
  bamFiles,
  bamNames = bamFiles,
  minMQS = 255,
  maxFrag = 500,
  pe = "none",
  restrict = "chr1",
  winWidth = 5000,
  col = inferno,
  genome,
  GCprob = TRUE,
  span = 0.1
)
}
\arguments{
\item{bamFiles}{The filenames (including full path) of the bam files containing the reads you want to count around the peak summits.}

\item{bamNames}{A name to describe the bam files you are using (for example: "H3K9me3_reads"). If no names are supplied, the full bamFiles strings are used.}

\item{minMQS}{The minimum mapping quality. Default is 255, which eliminates multimapping reads when STAR was used to generate bam files.}

\item{maxFrag}{An integer scalar, specifying the maximum fragment length corresponding to a read pair. Defaults to 500 base pairs.}

\item{pe}{A string indicating whether paired-end data is present; set to "none" (the default), "both", "first" or "second".}

\item{restrict}{A character vector containing the names of allowable chromosomes from which reads will be extracted. Default is "chr1".}

\item{winWidth}{An integer scalar specifying the width of the window, in which reads are counted and GC content calculated. Default is 5000 base pairs.}

\item{col}{Color scheme for the plots. Default is "plasma" from the viridis package.}

\item{genome}{A BSGenome object for the species you are analyzing. Required parameter. For example,use BSgenome.Mmusculus.UCSC.mm10 for mouse.}

\item{GCprob}{Should the GC content be displayed as absolute counts (GCprob=FALSE) or as fraction of GCs (GCprob=TRUE,default).}

\item{span}{Span that is used for loess trendline. Default= 0.1}
}
\value{
A smooth Scatter Plot for every suuplied alignment file: The number of Gs and Cs in each window is plotted on the x-axis, the read (fragment if pe="both") counts
(in counts/million) is plotted on the y-axis.
}
\description{
Plot GC content versus read counts.
}
\details{
This function generates a plot of the number of Gs and Cs against the read count (cpm) in windows across the chosen genome.
A seperate plot is generated for each alignment file (bam file) that is supplied. Supports both single and paired-end experiments.
}
\examples{
bamFiles <- c("/work2/gbuehler/deepSeqRepos/bam//HP1a_wt_ChIP_r1_818F1_multi.bam","/work2/gbuehler/deepSeqRepos/bam//HP1a_wt_ChIP_r2_818F3_multi.bam")
GCbias(bamFiles=bamFiles)

}
