% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log2Transform.R
\name{Log2Transform}
\alias{Log2Transform}
\title{Log2Transform}
\usage{
Log2Transform(counts, pseudocount = 1)
}
\arguments{
\item{counts}{A list of counts across peak summits generated by SummitHeatmap.}

\item{pseudocount}{The pseudocount to add to the values (to avoid taking the log2 of 0). Default=1.}
}
\value{
A list of matrices of same size as counts, with all the values log2 transformed.
}
\description{
This function takes a list of matrices and calculates the log2 for each value.
If (and only if) at least one value in a given matrix is 0, a pseudocount is added first to all values.
}
