% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniplot.R
\name{miniplot}
\alias{miniplot}
\title{miniplot}
\usage{
miniplot(number1, number2)
}
\arguments{
\item{number1}{the first number}

\item{number2}{the second number}
}
\value{
a plot
}
\description{
this makes a silly plot
}
\examples{
miniplot(1,2)

}
