% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{plotTracks}
\alias{plotTracks}
\title{plotTracks}
\usage{
plotTracks(trackData, bigwigNames, color_map)
}
\arguments{
\item{trackData}{A list of data frames generated with calcTracks.}

\item{bigwigNames}{Character vector containing the names to describe the \code{bigwigFiles} you are using (for example: "H3K9me3_replicate1").}

\item{color_map}{Named vector indicating the plotting color for each group of tracks.(for example: c("H3K9me3" = "red")).}
}
\value{
A plot of genomic tracks of the data in \code{bigwigFiles} and the repeat and gene annotation.
}
\description{
Plot tracks for a given genomic region of interest (ROI).
}
\details{
plotTracks is used to take the ChIP, repeat, and gene annotation data using \code{calcTracks} and plot tracks
for a given region of interest (ROI).
}
\examples{
#see vignette

}
