% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim_matrix.R
\name{redim_matrix}
\alias{redim_matrix}
\title{redim_matrix}
\usage{
redim_matrix(
  mat,
  target_height = 100,
  target_width = 100,
  summary_func = function(x) mean(x, na.rm = TRUE),
  output_type = 0,
  n_core = 1
)
}
\arguments{
\item{mat}{A list of counts across peak summits generated by SummitHeatmap.}

\item{target_height}{The desired number of rows of the output matrix.}

\item{target_width}{The desired number of columns of the output matrix.}

\item{summary_func}{The function to use for summarising the matrix.Default=mean.}

\item{output_type}{The type of output matrix.}

\item{n_core}{Numebr of cores to sue for calculation.}
}
\value{
A matrix of the size target_height by target_width that summarizes the original matrix.
}
\description{
This function summarizes a matrix into the desired numbe rof rows and columns.
}
\details{
It was taken from this reference: https://gdevailly.netlify.app/post/plotting-big-matrices-in-r/
}
