% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePeaks.R
\name{SimulatePeaks}
\alias{SimulatePeaks}
\title{SimulatePeaks}
\usage{
SimulatePeaks(nSites, peak.widths)
}
\arguments{
\item{nSites}{The number of ranges.}

\item{peak.widths}{A numeric vector of the length nSites containing the desired width of the ranges.}
}
\value{
A GRanges object of randomly chosen genomic ranges of length nSites with widths peak.widths.
}
\description{
This function generates a GRanges object with randomly chosen ranges
in the Mus musculus mm10 genome.
}
\details{
This function can be used to shuffle peak regions of your
ChIP data to random locations in the mouse genome. These random peaks
can be used as control regions for further analysis, for example, when
testing the overlap of your peaks with genomic annotations (genes, repeats,...).
}
\examples{
SimulatePeaks(1000,rep(100,1000))

}
