% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountPeakReads.R
\name{CountPeakReads}
\alias{CountPeakReads}
\title{CountPeakReads}
\usage{
CountPeakReads(
  peaks,
  bamFiles,
  bamNames,
  chips,
  inputs,
  width = 500,
  minOverlap = 1,
  PairedEnd = FALSE,
  minMQS = 255,
  strand = 0,
  splitOnly = FALSE,
  nonSplitOnly = FALSE,
  readExtension3 = 0,
  readShiftSize = 0,
  requireBothEndsMapped = FALSE,
  read2pos = 5,
  pseudocount = 8
)
}
\arguments{
\item{peaks}{A Granges object containing your positions of interest. Must include seqnames (chromosomes), start, end, strand, and name.}

\item{bamFiles}{The filenames (including full path) of the bam files containing the reads you want to count around the peak summits.}

\item{bamNames}{A name to describe the bam files you are using (for example: "H3K9me3_reads"). Required argument.}

\item{chips}{The names of the ChIP samples (corresponding exactly to the ChIP sample names in bamNames).}

\item{inputs}{The names of the Input samples (corresponding exactly to the Input sample names in bamNames),
provided in an order crresponding to the ChIP samples. If missing, no Input normlaization is performed.}

\item{width}{The width around the center of the peaks GRanges object provided in which the reads will be counted,default = 500.
If width=0, the provided regions will remain the original size.}

\item{minOverlap}{The minimum overlap of a read with the window for the read to be counted. Default = 1.}

\item{PairedEnd}{Are reads in bam files generated with paired-end or single-end sequencing. Default is FALSE (=single-end).}

\item{minMQS}{The minimum mapping quality. Default is 255, which eliminates multimapping reads when STAR was used to generate bam files.}

\item{strand}{Strand specific sequencing? If the reads in the bam file are derived from both strands (usually the case for ChIP experiments),
use 0 (default). 1 or 2 are the strans specific options (as in feature counts).}

\item{splitOnly}{Should only split reads (for example spliced reads) be counted? FALSE by default.}

\item{nonSplitOnly}{Should only non-split reads (for example not spliced reads) be counted? FALSE by default.}

\item{readExtension3}{integer giving the number of bases extended downstream from 3' end of each read. 0 by default. Negative value is not allowed.}

\item{readShiftSize}{integer specifying the number of bases the reads will be shifted downstream by. 100 by default. Negative value is not allowed.}

\item{requireBothEndsMapped}{logical indicating if both ends from the same fragment are required to be successfully aligned before the fragment can be assigned to a feature or meta-feature. This parameter is only appliable when isPairedEnd is TRUE.}

\item{read2pos}{Specifying whether each read should be reduced to its 5' most base or 3' most base. It has three possible values: NULL, 5 (denoting 5' most base) and 3 (denoting 3' most base).
Default value is 5, ie. the 5' end of the read will be counted. If a read is reduced to a single base, only that base will be considered for the read assignment.
Read reduction is performed after read shifting and extension.}

\item{pseudocount}{The pseudocount that is added to the counts of Chip and Input before normalization to avoind dividing by 0. Default is 8.}
}
\value{
A list of 2 matrices. Both with a column for each chip sample provided, and the same number of rows as regions provided in the GRanges object.
The first matrix contains the (Input) normalized number of reads in a window around each position of interest, which was the center in the supplied GRanges object.
The second matrix countains the total number of reads in the the chip sample, or in the chip plus the input sample, if an input sample is provided.
}
\description{
Generate normalized counts around positions of interest in a genome.
}
\details{
This function calculates the numebr of reads of different experiments around positions of interest in the genome,
for example the summits of ChIP peaks, or transcription start sites. The center of your GRanges object defines the coordinates
of these positions of interest (eg peak summits, TSSs). Then the number of reads in a ChIP sample is normalized to the number
of reads in the corresponding Input sample, if provided, taking the total library sizes into account.
If no Input sample names are provided, the counts per milliion reads (cpm) for each ChIP samples is returned.
}
\examples{
peaks <- GenomicRanges::GRanges(
seqnames = Rle(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
ranges = IRanges(50101:50110, end = 51111:51120),
strand = Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
summit = 1:10, name = head(letters, 10))
bamFiles <- list.files(system.file("extdata", package = "MiniChip"), full.names=TRUE,pattern="*bam$")
bamNames <- gsub(paste(system.file("extdata", package = "MiniChip"),"/",sep=""),"",bamFiles)
bamNames <- gsub("_chr11.bam","",bamNames)
CountPeakReads(peaks=peaks,bamFiles=bamFiles,bamNames=bamNames,chips=bamFiles[1:3],inputs=bamFiles[4:6])

}
