% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawSummitHeatmaps.R
\name{DrawSummitHeatmaps}
\alias{DrawSummitHeatmaps}
\title{DrawSummitHeatmaps}
\usage{
DrawSummitHeatmaps(counts, bamNames, nicebamNames = bamNames,
  plotcols = circlize::colorRamp2(c(0, 3), c("white", "darkblue")))
}
\arguments{
\item{counts}{A list of counts across peak summits generated by SummitHeatmap.}

\item{bamNames}{A character vector of names to describe the bam files you are using. This should match colnames(counts).}

\item{nicebamNames}{A character vector of short names to describe the bam files you are using. Needs to be the same length as bamNames.}

\item{plotcols}{The colors for the heatmap. Default is circlize::colorRamp2(c(0, 3), c("white", "darkblue")).}
}
\value{
A complex heatmap object containing a heatmap for each sample (each bam file),
sorted by the read counts in the middle window (around peak summit) of the first sample (bam file) that was provided.
}
\description{
This function plots heatmaps of pre-calculated read counts around the summits of ChIP peaks.
}
\details{
This function plots heatmaps of pre-calculated read counts around the summits of ChIP peaks.
}
\examples{
counts <- list(matrix(rnorm(21000,2,1),ncol=21,nrow=100,dimnames=list(1:100,-10:10)),
              matrix(rnorm(21000,2,1),ncol=21,nrow=100,dimnames=list(1:100,-10:10)))
bamNames <- c("counts1","counts2")
names(counts) <- bamNames
DrawSummitHeatmaps(counts,bamNames)

}
