% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTracks.R
\name{plotTracks}
\alias{plotTracks}
\title{plotTracks}
\usage{
plotTracks(
  bwFilesPlus,
  bwFilesMinus,
  bwNames,
  txdb,
  EnsDb,
  bedFiles,
  gene,
  plotregion,
  plotExtension = 1000,
  plotranges = rep(1, length(bwNames))
)
}
\arguments{
\item{bwFilesPlus}{Character vector containing the filenames (including the full path) of read alignment files in bigwig format.
Either the only bigwig file per sample if unstranded (eg for ChIPseq) or the bigwig file containing the reads mapped on the plus strand.}

\item{bwFilesMinus}{Character vector containing the filenames (including the full path) of read alignment files in bigwig format,
containing the reads mapped on the minus strand. If missing, only the \code{bwFilesPlus} files will be used.}

\item{bwNames}{Character vector containing the names to describe the samples (for example: "H3K9me3").}

\item{txdb}{TxDb object containing all transcripts/genes using Ensembl Gene IDs.}

\item{EnsDb}{EnsDb object containing an Ensembl annotation database linking GeneIDs to gene symbols.}

\item{bedFiles}{Optional character vector containing the filenames (including the full path) of any bed files (eg peaks) to be plotted.
They should contain columns: chrom, start,stop, gene, score, strand (-1 or 1),type (exon).}

\item{gene}{Character scalar specifying the gene symbol for the gene around which the tracks should be plotted.
If missing, a \code{plotregion} needs to be supplied instead.}

\item{plotregion}{A GRanges object containing the region of the genome for plotting the tracks (in case \code{gene} is not supplied).}

\item{plotExtension}{Integer scalar specifying the length (in bp) up and downstream of the gene the plot should be extended. Default = 1000.}

\item{plotranges}{Numeric vector of the same length as bwNames specifying the y-axis range for each plot. Default=1 for each sample.}
}
\value{
A plot of genome browser like tracks, including transcripts.
}
\description{
Plots genome browser like tracks.
}
\details{
Plots genome browser like tracks for supplied data (stranded or unstranded, eg RNAseq, CHIPseq) in bigwig format,
for a region surrounding a specified gene, or a region specified by a GRanges object. Including a track for all transcripts in that region.
}
\examples{
library(TxDb.Mmusculus.UCSC.mm10.ensGene)
library(EnsDb.Mmusculus.v79)
bwFiles <- list.files(system.file("extdata", package = "MiniChip"), full.names=TRUE,pattern="*bw$")
plotTracks(bwFilesPlus=bwFiles,bwNames=c(rep("Adnp",2)), txdb=TxDb.Mmusculus.UCSC.mm10.ensGene,EnsDb=EnsDb.Mmusculus.v79,
gene="Adnp", plotExtension=500,plotranges=rep(0.8,2))
}
