% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeHeatmaps.R
\name{SummarizeHeatmaps}
\alias{SummarizeHeatmaps}
\title{SummarizeHeatmaps}
\usage{
SummarizeHeatmaps(counts, sampleList, summarizing = mean)
}
\arguments{
\item{counts}{A list of counts across peak summits generated by SummitHeatmap.}

\item{sampleList}{A list where each element is a character vector of sample names that should be grouped.
The name of each element of this list specifies the group name.}

\item{summarizing}{Defines the way in which the counts are summarized. Defaults to mean.}
}
\value{
A list of summarized (by group) counts across peak summits.
}
\description{
Summarize heatmap count matrices from individual bam samples.
}
\details{
This function takes a list of heatmap count matrices (from individual bam files) and summarizes them (by user defined groups).
The matrices should have unique row names.
}
\examples{
counts <- list(matrix(rnorm(21000,2,1),ncol=21,
nrow=100,dimnames=list(1:100,-10:10)),
               matrix(rnorm(21000,2,1),ncol=21,
               nrow=100,dimnames=list(1:100,-10:10)),
               matrix(rnorm(21000,2,1),ncol=21,
               nrow=100,dimnames=list(1:100,-10:10)),
               matrix(rnorm(21000,2,1),ncol=21,
               nrow=100,dimnames=list(1:100,-10:10)))
bamNames <- c("counts1","counts2","counts3","counts4")
names(counts) <- bamNames
SummarizeHeatmaps(counts,sampleList=list
(countsA=c("counts1","counts2"),countsB=c("counts3","counts4")))

}
