% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummitHeatmap.R
\name{SummitHeatmap}
\alias{SummitHeatmap}
\title{SummitHeatmap}
\usage{
SummitHeatmap(peaks, peaknames = "mypeaks", bamFiles,
  bamNames = "myreads", span = 2025, step = 50,
  plotcols = colorRampPalette(brewer.pal(9, "RdPu"))(255),
  plotHM = TRUE, useCPM = TRUE, PairedEnd = FALSE)
}
\arguments{
\item{peaks}{A Granges object containing your ChIP peaks. Must include seqnames (chromosomes), start, end, strand, summit, and name.}

\item{peaknames}{A character string that describes your peaks. Default is "mypeaks".}

\item{bamFiles}{The filenames (including full path) of the bam files containing the reads you want to count around the peak summits.}

\item{bamNames}{A name to describe the bam files you are using (for example: "H3K9me3_reads"). Defaults to "myreads".}

\item{span}{The distance from the peak summit to the left and right that you want your heatmap to extand to. Default is 2025.}

\item{step}{The window size in which reads are counted/plotted in the heatmap. Default is 50.}

\item{plotcols}{The colors for the heatmap. Default is colorRampPalette(brewer.pal(9,"RdPu"))(255).}

\item{plotHM}{Also plot the heatmap or only count the reads in the windows. Default is TRUE.}

\item{useCPM}{Normalize the number of reads per window to the total number of reads in the sample (bam file). Default is TRUE.}

\item{PairedEnd}{Are reads in bam files generated with paired-end or single-end sequencing. Default is FALSE (=single-end).}
}
\value{
A list of matrices of the same length as the number of bam files supplied.
Each matrix contains the number of reads (or cpm) in each window (column headers=middle of the window)
for each peak (rownames=peaknames). If plotHM is TRUE, a heatmap is plotted for each sample as well (each bam file)
and saved in a directory (in your current working directory) that is named after your peaknames object.
These heatmaps are clustered individually.
}
\description{
This function generates a heatmap of different experiments (read counts) around the summits of ChIP peaks.
}
\details{
This function.
}
\examples{
peaks <- GRanges(
seqnames = Rle(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
ranges = IRanges(50101:50110, end = 51111:51120),
strand = Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
summit = 1:10, name = head(letters, 10))
bamFiles <- c("/work2/gbuehler/deepSeqRepos/bam//HP1a_wt_ChIP_r1_818F1_multi.bam","/work2/gbuehler/deepSeqRepos/bam//HP1a_wt_ChIP_r2_818F3_multi.bam")
SummitHeatmap(peaks=peaks,bamFiles=bamFiles,plotHM=FALSE)

}
